0.1.6 - Breaking and Combining along with more
- Down arrow now opens the breaking and combining menu (yay!). The left
side is what is on your vest and backpack. If it is red, there are no
possible output for it. If it is white, there is!. On the left side is
the possible output for the selected input. If I have enough input, the
output selection will be white and have a number displaying how many I
can create. If I do not have enough, it will be red and say 0.
- cfgLoot now spawns the correct mags for P07 and Rook.
- Group menu no longer has weird text clipping. I will create my own
group menu in the future.
- Fixed bugs and typos.